Add logic for handling game input from the player
Add logic for taking input based on user clicking the desired weapon of choice.
Add logic for updating score and rounds on the screen.
Add logic for determining win, lose, and tie conditions
Add other pages for redirecting the player if he either wins or loses
Add logic to reload the page in case a tie happens and there are no rounds left